+ fixed a movement data packet sending ( maybe it fix the client error on transport teleportation  ) + fixed a typo on creature displayinfo structure ( it will fix client error on login ) + deleted one TBC go script from SD2 proj.

+++ src/shared/Database/DBCStructure.h
@@ -139,8 +139,7 @@ struct CreatureDisplayInfoEntry
     uint32      Displayid;                                  // 0
 	                                                        // 1	iRefID_CreatureModelData Points to the model filename.
                                                             // 3	iRefID_CreatureDisplayInfoExtra	
-                                                            // 4	
-    int32       scale;                                      // 4    Scale  Default model scale if not set by server for that spawn
+    float       scale;                                      // 4    Default model scale if not set by server for that spawn
                                                             // 6	    Opacity	 Integer	 0 - 255. 0 is totally transparent, 255 is solid	
 };
